feat: update task dto and implement handler to not found task id

+++ back-end-nestjs/src/tasks/tasks.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from 'src/database/prisma.service';
+import { TasksRepository } from './repository/task-repository';
 
 @Injectable()
-export class TasksService {
+export class TasksService implements TasksRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   async save(task: CreateTaskDto) {
@@ -32,10 +33,16 @@ export class TasksService {
   }
 
   async findOne(id: number) {
+    const taskExists = await this.verifyTaskExists(id);
+    if (!taskExists) throw new NotFoundException('Task not found');
+
     return await this.prisma.task.findUnique({ where: { id } });
   }
 
   async update(id: number, updateTaskDto: UpdateTaskDto) {
+    const taskExists = await this.verifyTaskExists(id);
+    if (!taskExists) throw new NotFoundException('Task not found');
+
     return await this.prisma.task.update({
       where: { id },
       data: {
@@ -57,6 +64,15 @@ export class TasksService {
   }
 
   async remove(id: number) {
+    const taskExists = await this.verifyTaskExists(id);
+    if (!taskExists) throw new NotFoundException('Task not found');
+
     return await this.prisma.task.delete({ where: { id } });
+  }
+
+  async verifyTaskExists(id: number): Promise<boolean> {
+    const verifyTask = await this.prisma.task.findUnique({ where: { id } });
+
+    return verifyTask !== null;
   }
 }
